Forgot to stage files, add test files
