+++ test/doBackwardsSearch.test.js
@@ -2,50 +2,7 @@ const assert = require('assert');
 const { ReferenceProvider } = require('../extension');
 const referenceProvider = new ReferenceProvider();
 
-suite('ReferenceProvider', () => {
-
-    test('getModulesWithPathFromRequireOrDefine should return object with module path and name', () => {
-        const input = `define(['./path/to/a', './path/to/b'], function (moduleA, moduleB) {`;
-        const expected = {
-            'moduleA': './path/to/a',
-            'moduleB': './path/to/b'
-        };
-        
-        assert.deepEqual(referenceProvider.getModulesWithPathFromRequireOrDefine(input), expected);
-    });
-
-    test('getRequireOrDefineStatement should return statement', () => {
-        const input = `
-            define(['./module'], function (module) {
-            })
-        `;
-        const expected = `define(['./module'], function (module`;
-        
-        assert.equal(referenceProvider.getRequireOrDefineStatement(input), expected);
-    });
-
-    test('removeComments should remove comments from text', () => {
-        const input = `
-            define(['./module'], function (module) {
-                // require();
-                /* comment */
-                /**
-                 * comment
-                 */
-            })
-        `;
-        const forceWhiteSpace = '';
-        const expected = `
-            define(['./module'], function (module) {
-                ${forceWhiteSpace}
-                ${forceWhiteSpace}
-                ${forceWhiteSpace}
-            })
-        `;
-        
-        assert.equal(referenceProvider.removeComments(input), expected);
-    });
-
+suite('doBackwardsSearch', () => {
     test('doBackwardsSearch returns offset of character - 1 if character at offset equals searchFor', () => {
         const input = 'foo.baz();';
         const input2 = `require('moduleA').foo();`;
@@ -81,5 +38,5 @@ suite('ReferenceProvider', () => {
         const expected = -1;
 
         assert.equal(referenceProvider.doBackwardsSearch(input, input.length-1, '.'), expected);
-    });
+    }); 
 });
